Convert the value to g/100 unit
When the sensor is placed perfectly leveled with the ground, the value
of x/y is 0, and the value of z is +16384 at a sensitivity of 2g.  In
order to get the real value, we have to multiplied it by 1g/16384.
I divided it by 164 to make it in g/100 unit